fix afk error

+++ alita/plugins/afk.py
@@ -118,7 +118,10 @@ async def rem_afk(c: Alita, m: Message):
         await m.reply_text(f"Error while checking afk\n{ef}")
         return
 
-    since = strftime("%Hh %Mm %Ss", gmtime(time() - user_afk["time"]))
+    if not user_afk:
+        return
+
+    since = strftime("%Hh %Mm %Ss", gmtime(time() - (user_afk["time"])))
     await db.remove_afk(m.from_user.id)
     user = await c.get_users(user_afk["user_id"])
     await m.reply_text(
